+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.5.2 (Jun 30, 2019)
 
-- Alway add a timestamp to the url of a cross origin image (#519).
+- Alway add a timestamp to the URL of a cross-origin image (#519).
 
 ## 1.5.1 (Mar 10, 2019)