+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## next
+## 1.5.3 (Jul 10, 2019)
 
 - Compute the size of image specially for all browsers that use WebKit as the layout engine in iOS devices, such as Safari and in-app browsers (#544).